requirements.txt
discord.py==2.5.2
python-dotenv==1.1.1
pytz==2025.2
aiohttp==3.12.15